refactor: split eligibility modal into smaller components
