+++ src/Widgets/EligibilityModal/index.tsx
@@ -1,13 +1,12 @@
 import LogoIcon from 'assets/Logo'
-import cx from 'classnames'
 import Modal from 'components/Modal'
-import { isToday } from 'date-fns'
 import React, { FunctionComponent, useState } from 'react'
-import { FormattedDate, FormattedMessage, FormattedNumber, useIntl } from 'react-intl'
+import { FormattedMessage } from 'react-intl'
 import { EligibilityPlan } from 'types'
-import { priceFromCents } from 'utils'
-import { paymentPlanShorthandName } from 'utils/paymentPlanStrings'
 import s from './EligibilityModal.module.css'
+import EligibilityPlans from './EligibilityPlans'
+import Info from './Info'
+import Installments from './Installments'
 
 type Props = {
   initialPlanIndex?: number
@@ -19,17 +18,8 @@ const EligibilityModal: FunctionComponent<Props> = ({
   onClose,
   eligibilityPlans,
 }) => {
-  const intl = useIntl()
   const [currentPlanIndex, setCurrentPlanIndex] = useState(initialPlanIndex || 0)
-
   const currentPlan = eligibilityPlans[currentPlanIndex]
-  const total =
-    currentPlan &&
-    priceFromCents(currentPlan.purchase_amount + currentPlan.customer_total_cost_amount)
-  const creditCost = currentPlan ? priceFromCents(currentPlan.customer_total_cost_amount) : 0
-  const TAEG = currentPlan?.annual_interest_rate && currentPlan.annual_interest_rate / 10000
-  const customerFees = priceFromCents(currentPlan ? currentPlan.customer_total_cost_amount : 0)
-  const isCredit = currentPlan && currentPlan.installments_count > 4
 
   return (
     <Modal onClose={onClose} ariaHideApp={false} scrollable isOpen>
@@ -41,125 +31,13 @@ const EligibilityModal: FunctionComponent<Props> = ({
             values={{ highlighted: (...chunks: string[]) => <span>{chunks}</span> }}
           />
         </div>
-        <div className={s.buttons}>
-          {eligibilityPlans.map((eligibilityPlan, index) => (
-            <button
-              key={index}
-              className={cx({ [s.active]: index === currentPlanIndex })}
-              onClick={() => setCurrentPlanIndex(index)}
-            >
-              {paymentPlanShorthandName(eligibilityPlan)}
-            </button>
-          ))}
-        </div>
-        <div className={s.schedule} data-testid="modal-installments-element">
-          <div className={cx(s.scheduleLine, s.total)}>
-            <span>
-              <FormattedMessage id="eligibility-modal.total" defaultMessage="Total" />
-            </span>
-            <span>
-              <FormattedNumber value={total} style="currency" currency="EUR" />
-            </span>
-          </div>
-          <div className={cx(s.scheduleLine, s.creditCost)}>
-            {isCredit ? (
-              <span>
-                <FormattedMessage
-                  id="eligibility-modal.credit-cost"
-                  defaultMessage="Dont coût du crédit"
-                />
-              </span>
-            ) : (
-              <span>
-                <FormattedMessage id="eligibility-modal.cost" defaultMessage="Dont frais" />
-              </span>
-            )}
-            <span>
-              {isCredit ? (
-                <FormattedMessage
-                  id="eligibility-modal.credit-cost-amount"
-                  defaultMessage="{creditCost} (TAEG {TAEG})"
-                  values={{
-                    creditCost: intl.formatNumber(creditCost, {
-                      style: 'currency',
-                      currency: 'EUR',
-                    }),
-                    TAEG: intl.formatNumber(TAEG ?? 0, {
-                      style: 'percent',
-                      maximumFractionDigits: 2,
-                    }),
-                  }}
-                />
-              ) : (
-                <FormattedNumber value={customerFees} style="currency" currency="EUR" />
-              )}
-            </span>
-          </div>
-          {(currentPlan?.payment_plan || []).map((installment, index) => (
-            <div className={s.scheduleLine} key={index}>
-              <span>
-                {isToday(installment.due_date * 1000) ? (
-                  <FormattedMessage id="installments.today" defaultMessage="Aujourd'hui" />
-                ) : (
-                  <FormattedDate
-                    value={installment.due_date * 1000}
-                    day="numeric"
-                    month="long"
-                    year="numeric"
-                  />
-                )}
-              </span>
-              <span>
-                <FormattedNumber
-                  value={priceFromCents(installment.total_amount)}
-                  style="currency"
-                  currency="EUR"
-                />
-              </span>
-            </div>
-          ))}
-          {isCredit && (
-            <p className={s.creditMessage}>
-              <FormattedMessage
-                id="eligibility-modal.credit-commitment"
-                defaultMessage="Un crédit vous engage et doit être remboursé. Vérifiez vos capacités de remboursement
-              avant de vous engager."
-              />
-            </p>
-          )}
-        </div>
-        <div className={s.list} data-testid="modal-info-element">
-          <div className={s.listItem}>
-            <div className={s.bullet}>1</div>
-            <div>
-              <FormattedMessage
-                id="eligibility-modal.bullet-1"
-                defaultMessage="Choisissez <strong>Alma</strong> au moment du paiement."
-                values={{ strong: (...chunks: string[]) => <strong>{chunks}</strong> }}
-              />
-            </div>
-          </div>
-          <div className={s.listItem}>
-            <div className={s.bullet}>2</div>
-            <div>
-              <FormattedMessage
-                id="eligibility-modal.bullet-2"
-                defaultMessage="Renseignez les <strong>informations</strong> demandées."
-                values={{ strong: (...chunks: string[]) => <strong>{chunks}</strong> }}
-              />
-            </div>
-          </div>
-          <div className={s.listItem}>
-            <div className={s.bullet}>3</div>
-            <div>
-              <FormattedMessage
-                id="eligibility-modal.bullet-3"
-                defaultMessage="La validation de votre paiement <strong>instantanée</strong> !"
-                values={{ strong: (...chunks: string[]) => <strong>{chunks}</strong> }}
-              />
-            </div>
-          </div>
-        </div>
+        <EligibilityPlans
+          eligibilityPlans={eligibilityPlans}
+          currentPlanIndex={currentPlanIndex}
+          setCurrentPlanIndex={setCurrentPlanIndex}
+        />
+        <Installments currentPlan={currentPlan} />
+        <Info />
         <div className={s.logo}>
           <LogoIcon underlineColor="#FF414D" />
         </div>
